requirements.txt
altgraph==0.17
Babel==2.9.1
click==7.1.2
cmake==3.18.4.post1
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
tkcalendar==1.6.1
